fix: 🐛 .md in the menu and fixed the warnings in the build
