+++ lib/node.js
@@ -2,13 +2,8 @@ import path from 'path'
 import fs from 'fs'
 
 export const Node = {
-  isFile: function (filename) {
-    try {
-      return fs.lstatSync(filename).isFile()
-    } catch (err) {
-      console.log(err)
-      return false
-    }
+  fileExists: function (filename) {
+    return fs.existsSync(filename)
   },
   getFullPath: function (folderPath) {
     return fs.readdirSync(folderPath).map(fn => path.join(folderPath, fn))
