+++ lib/utils.js
@@ -87,7 +87,7 @@ export function getSlugHashMap () {
 }
 
 export function toSlug (filePath) {
-  if (Node.isFile(filePath) && filePath.includes(Node.getMarkdownFolder())) {
+  if (Node.fileExists(filePath) && filePath.includes(Node.getMarkdownFolder())) {
     return filePath.replace(Node.getMarkdownFolder(), '')
       .replaceAll('/', '__')
       .replaceAll(' ', '++++')
@@ -101,7 +101,7 @@ export function toSlug (filePath) {
 
 export function constructGraphData () {
   const filepath = path.join(process.cwd(), 'graph-data.json')
-  if (Node.isFile(filepath)) {
+  if (Node.fileExists(filepath)) {
     const data = fs.readFileSync(filepath)
     return JSON.parse(String(data))
   } else {
@@ -223,6 +223,7 @@ export function convertObject (thisObject) {
   routerPath = routerPath ? '/note/' + routerPath : null
   const newObject = {
     name: thisObject.name,
+    label: thisObject.name.replace('.md', ''),
     children,
     id: (_counter++).toString(),
     routePath: routerPath || null
@@ -238,22 +239,6 @@ export function convertObject (thisObject) {
     return newObject
   }
 }
-
-function flat (array) {
-  let result = []
-  array.forEach(function (a) {
-    result.push(a)
-    if (Array.isArray(a.children)) {
-      result = result.concat(flat(a.children))
-    }
-  })
-  return result
-}
-
-export function getFlattenArray (thisObject) {
-  return flat(thisObject.children)
-}
-
 export function commaAndList (aList) {
   switch (aList.length) {
     case 0:
